Tmp Debugging
There's an issue with the creating of the file in the /tmp directory, I'm not completely sure why it isn't working.

https://stackoverflow.com/questions/4632028/how-to-create-a-temporary-directory
https://unix.stackexchange.com/questions/181937/how-create-a-temporary-file-in-shell-script
https://github.com/0mp/vipe.sh/blob/master/vipe.sh
https://stackoverflow.com/questions/40362369/stdin-read-fails-on-some-input

I also rediscovered an issue I've had before with NBTify's CLI, in that you can't pipe to and from itself with multiple calls, say like `echo { hey: 5 }` | nbtify --nbt | nbtify --snbt`. I think it might be something to do with loading the content from `process.stdin.fd`, maybe because I'm doing it synchronously as well, I'm not completely sure.